backend/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn==0.25.0
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Environment variables
python-dotenv==1.0.0

# Testing
pytest==7.4.4
httpx==0.26.0
pytest-asyncio==0.23.2

# Type checking
mypy==1.8.0